New material ui

+++ ShikkhanobishStudentApp/ShikkhanobishStudentApp/Model/Institution.cs
@@ -10,7 +10,7 @@ namespace ShikkhanobishStudentApp.Model
         public string title { get; set; }
         public string name { get; set; }
         public int tuitionRequest { get; set; }
-        public double avgRatting { get; set; }
+        public float avgRatting { get; set; }
         public int indexNo { get; set; }
         public string Response { get; set; }
